Add files via upload
train.csv - A file with a table of houses, their specifications, and their individual prices. <<<Will be used to find the correlation between specifications and price.

test.csv - Similar to train, but without the prices of each house. <<< Will be used to calculate the prices based on correlations found from the train file.

data_description.txt - A text file that explains what each specification means, and, for some, their possible values.