Add files via upload
Removed code that I believed would be unfair to label as my own. 

Once the code was removed, it dropped the score to 0.12786, from 0.12515
By refining my own code, I had raised the score to 0.12453. Yielded a rank of 1343.

Could potentially be used as a final submission, as the code no longer contains lines that weren't referenced.
Additionally, many arguements are now only similar to their sources, but not exact. Will require further guidance as to what is and isn't plagiarizing in matters of code.